// apps/annie/src/client/commands/random-faction/types.ts
export enum FactionType {
  BandleCity = "Bandle City",
  Bilgewater = "Bilgewater",
  Demacia = "Demacia",
  Freljord = "Freljord",
  Ionia = "Ionia",
  Ixtal = "Ixtal",
  Noxus = "Noxus",
  Piltover = "Piltover",
  ShadowIsles = "Shadow Isles",
  Shurima = "Shurima",
  Targon = "Targon",
  Void = "Void",
  Zaun = "Zaun",
}

export interface Faction {
  name: FactionType;
  imgUrl: string;
  challenge: string;
  champions: string[];
}